Break four independent import cycles by extracting leaf modules and context hooks
