+++ src/components/ConnectToCertiniaFlow/index.tsx
@@ -9,7 +9,7 @@ import usePopoverPosition from '@hooks/usePopoverPosition';
 import {isAuthenticationError} from '@libs/actions/connections';
 import Navigation from '@libs/Navigation/Navigation';
 
-import {useAccountingState} from '@pages/workspace/accounting/AccountingContext';
+import {useAccountingState} from '@pages/workspace/accounting/AccountingContext/contexts';
 
 import type {AnchorPosition} from '@styles/index';
