+++ src/components/ConnectToQuickbooksOnlineFlow/BaseConnectToQuickbooksOnlineFlow.tsx
@@ -12,7 +12,7 @@ import usePopoverPosition from '@hooks/usePopoverPosition';
 
 import {isAuthenticationError} from '@libs/actions/connections';
 
-import {useAccountingState} from '@pages/workspace/accounting/AccountingContext';
+import {useAccountingState} from '@pages/workspace/accounting/AccountingContext/contexts';
 
 import type {AnchorPosition} from '@styles/index';
