+++ src/libs/EmojiUtils.tsx
@@ -22,6 +22,7 @@ import React from 'react';
 import type {getEmojiTrie as getEmojiTrieType} from './EmojiTrie';
 
 import {isSafari} from './Browser';
+import {containsCustomEmoji, containsOnlyCustomEmoji} from './CustomEmojiUtils';
 import memoize from './memoize';
 
 type HeaderIndices = {code: string; index: number; icon: IconAsset};
@@ -871,24 +872,6 @@ function getProcessedText(processedTextArray: TextWithEmoji[], style: StyleProp<
             text
         ),
     );
-}
-
-function containsCustomEmoji(text?: string): boolean {
-    if (!text) {
-        return false;
-    }
-
-    const privateUseAreaRegex = CONST.REGEX.PRIVATE_USER_AREA;
-    return privateUseAreaRegex.test(text);
-}
-
-function containsOnlyCustomEmoji(text?: string): boolean {
-    if (!text) {
-        return false;
-    }
-
-    const privateUseAreaRegex = CONST.REGEX.ONLY_PRIVATE_USER_AREA;
-    return privateUseAreaRegex.test(text);
 }
 
 /**
