+++ src/libs/GPSDraftDetailsUtils.ts
@@ -6,13 +6,13 @@ import {stopGpsTripNotification} from '@pages/iou/request/step/IOURequestStepDis
 import type {GpsDraftDetails} from '@src/types/onyx';
 import type {GPSPoint, TrimmedGPSPoint} from '@src/types/onyx/GpsDraftDetails';
 import type {Routes, Waypoint} from '@src/types/onyx/Transaction';
-import geodesicDistance from '@src/utils/geodesicDistance';
 
 import type {SetRequired} from 'type-fest';
 
-import {hasStartedLocationUpdatesAsync, reverseGeocodeAsync, stopLocationUpdatesAsync} from 'expo-location';
+import {hasStartedLocationUpdatesAsync, stopLocationUpdatesAsync} from 'expo-location';
 
 import {removeLastSegment, setEndWaypointAddress, setIsTracking} from './actions/GPSDraftDetails';
+import {addressFromGpsPoint, calculateTrimmedEndPoint, coordinatesToString} from './GPSPointUtils';
 import {roundToTwoDecimalPlaces} from './NumberUtils';
 
 type GPSWaypointCollection = Record<string, SetRequired<Waypoint, 'keyForList' | 'lat' | 'lng' | 'address'>>;
@@ -115,27 +115,6 @@ function getStringifiedGPSCoordinates(gpsDraftDetails: GpsDraftDetails | undefin
     return JSON.stringify(updatedGpsPoints.map((points) => points.map(({lat, long}) => ({lng: long, lat}))));
 }
 
-async function addressFromGpsPoint(gpsPoint: {lat: number; long: number}): Promise<string | null> {
-    try {
-        const [location] = await reverseGeocodeAsync({latitude: gpsPoint.lat, longitude: gpsPoint.long});
-
-        if (!location) {
-            return null;
-        }
-
-        const address: string = location?.formattedAddress ?? [location?.name, location?.city, location?.region].filter(Boolean).join(', ');
-
-        return address;
-    } catch (error) {
-        console.error('[GPS distance request] Failed to reverse geocode location to postal address: ', error);
-        return null;
-    }
-}
-
-function coordinatesToString(gpsPoint: {lat: number; long: number}): string {
-    return `${gpsPoint.lat},${gpsPoint.long}`;
-}
-
 function isLastSegmentEmptyOrHasOnlyOnePoint(lastSegment: GPSPoint[]): boolean {
     if (lastSegment.length <= 1) {
         return true;
@@ -214,42 +193,6 @@ function getGpsPoints(gpsDraftDetails: GpsDraftDetails | undefined): GPSPoint[][
 
 function getFirstGpsPoint(gpsDraftDetails: GpsDraftDetails | undefined): GPSPoint | undefined {
     return gpsDraftDetails?.gpsPoints?.at(0)?.at(0);
-}
-
-function calculateTrimmedEndPoint(gpsPoints: GPSPoint[][], targetDistanceMeters: number): TrimmedGPSPoint | null {
-    let distanceTraveled = 0;
-
-    for (let segmentIndex = 0; segmentIndex < gpsPoints.length; segmentIndex++) {
-        const segment = gpsPoints.at(segmentIndex);
-
-        if (!segment) {
-            continue;
-        }
-
-        for (let pointIndex = 1; pointIndex < segment.length; pointIndex++) {
-            const previousPoint = segment.at(pointIndex - 1);
-            const currentPoint = segment.at(pointIndex);
-
-            if (!previousPoint || !currentPoint) {
-                continue;
-            }
-            const distanceBetweenPoints = geodesicDistance(previousPoint, currentPoint);
-
-            if (distanceTraveled + distanceBetweenPoints >= targetDistanceMeters) {
-                const fractionToInclude = distanceBetweenPoints === 0 ? 0 : (targetDistanceMeters - distanceTraveled) / distanceBetweenPoints;
-                const interpolatedPoint = {
-                    lat: previousPoint.lat + fractionToInclude * (currentPoint.lat - previousPoint.lat),
-                    long: previousPoint.long + fractionToInclude * (currentPoint.long - previousPoint.long),
-                };
-
-                return {...interpolatedPoint, segmentIndex, precedingPointIndex: pointIndex - 1};
-            }
-
-            distanceTraveled += distanceBetweenPoints;
-        }
-    }
-
-    return null;
 }
 
 function getTrimmedGpsTrip(gpsDraftDetails: GpsDraftDetails | undefined, trimmedEndPoint?: TrimmedGPSPoint): GPSPoint[][];
